+++ src/test/java/org/ays/auth/controller/AysAuthControllerTest.java
@@ -17,9 +17,9 @@ import org.ays.auth.model.response.AysTokenResponse;
 import org.ays.auth.service.AysAuthService;
 import org.ays.auth.service.AysUserPasswordService;
 import org.ays.common.model.response.AysErrorResponse;
+import org.ays.common.model.response.AysErrorResponseBuilder;
 import org.ays.common.model.response.AysResponse;
 import org.ays.common.model.response.AysResponseBuilder;
-import org.ays.common.util.exception.model.response.AysErrorResponseBuilder;
 import org.ays.util.AysMockMvcRequestBuilders;
 import org.ays.util.AysMockResultMatchersBuilders;
 import org.ays.util.AysValidTestData;
