+++ src/test/java/org/ays/auth/controller/AysPermissionControllerTest.java
@@ -7,8 +7,8 @@ import org.ays.auth.model.mapper.AysPermissionToPermissionsResponseMapper;
 import org.ays.auth.model.response.AysPermissionsResponse;
 import org.ays.auth.service.AysPermissionService;
 import org.ays.common.model.response.AysErrorResponse;
+import org.ays.common.model.response.AysErrorResponseBuilder;
 import org.ays.common.model.response.AysResponse;
-import org.ays.common.util.exception.model.response.AysErrorResponseBuilder;
 import org.ays.util.AysMockMvcRequestBuilders;
 import org.ays.util.AysMockResultMatchersBuilders;
 import org.junit.jupiter.api.Test;
