+++ src/test/java/org/ays/institution/controller/InstitutionControllerTest.java
@@ -2,8 +2,8 @@ package org.ays.institution.controller;
 
 import org.ays.AysRestControllerTest;
 import org.ays.common.model.response.AysErrorResponse;
+import org.ays.common.model.response.AysErrorResponseBuilder;
 import org.ays.common.model.response.AysResponse;
-import org.ays.common.util.exception.model.response.AysErrorResponseBuilder;
 import org.ays.institution.model.Institution;
 import org.ays.institution.model.InstitutionBuilder;
 import org.ays.institution.model.mapper.InstitutionToInstitutionsSummaryResponseMapper;
